+++ router/article.js
@@ -9,6 +9,7 @@ const multer = require("multer")
 const upload = multer({ dest: "uploadFiles/" })
 const fs = require("fs")
 const { verifyUser, checkPermission } = require("./middlewares/authorization")
+const paging = require("./js/pagination")
 
 // Status Code
 // 400 Bad Request
@@ -84,26 +85,6 @@ router.post("/", verifyUser, upload.array("fileList"), async (req, res) => {
         res.status(500).send({ message: "Server Error" })
     }
 })
-
-export const paging = (page, totalPost, limit) => {
-    page = parseInt(page)
-    totalPost = parseInt(totalPost)
-    limit = parseInt(limit)
-
-    let pageNum = page || 1
-    let postLimit = limit || 20
-    const pagination = 10
-    const hidePost = page === 1 ? 0 : (page - 1) * postLimit
-    const totalPages = Math.ceil((totalPost - 1) / postLimit)
-    const startPage = Math.floor((pageNum - 1) / pagination) * pagination + 1
-    let endPage = startPage + pagination - 1
-
-    if (pageNum > totalPages) pageNum = totalPages
-    if (endPage > totalPages) endPage = totalPages
-    if (totalPages === 0) endPage = 1
-
-    return { startPage, endPage, hidePost, postLimit, totalPages, pageNum }
-}
 
 router.delete("/delete", async(req, res) => {
     await Article.deleteMany({author: "hyejin"})
